Applied patch from Gargoyle project

+++ interpreter/parse.c
@@ -598,13 +598,13 @@ static void complexReferencesParser(ParameterPosition *parameterPosition) {
 static Bool restrictionCheck(RestrictionEntry *restriction, int instance) {
     if (restriction->class == RESTRICTIONCLASS_CONTAINER) {
         if (sectionTraceOption)
-            printf("\n<SYNTAX RESTRICTION WHERE parameter #%ld Isa Container, %s>\n",
+            printf("\n<SYNTAX RESTRICTION WHERE parameter #%d Isa Container, %s>\n",
                    restriction->parameterNumber,
                    instances[instance].container != 0?"PASSED":"FAILED:");
         return instances[instance].container != 0;
     } else {
         if (sectionTraceOption)
-            printf("\n<SYNTAX RESTRICTION WHERE parameter #%ld Isa %s[%ld], %s>\n",
+            printf("\n<SYNTAX RESTRICTION WHERE parameter #%d Isa %s[%d], %s>\n",
                    restriction->parameterNumber, idOfClass(restriction->class), restriction->class,
                    isA(instance, restriction->class)?"PASSED":"FAILED:");
         return isA(instance, restriction->class);
@@ -849,7 +849,7 @@ static void checkRestrictedParameters(ParameterPosition parameterPositions[], El
                         /* It wasn't ALL, we need to say something about it, so
                          * prepare a printout with $1/2/3
                          */
-                        sprintf(marker, "($%ld)", restriction->parameterNumber);
+                        sprintf(marker, "($%ld)", (unsigned long) restriction->parameterNumber);
                         setParameters(localParameters);
                         output(marker);
                         runRestriction(restriction, localParameters);
